WI00949199 - Fix Menu popup event not firing in .NET 8.0 (#30)
Form and MainMenu are tightly coupled in Winforms for .NET 4.8.

When we ported MainMenu to .NET 8 we missed some places in Form that also needed to be ported.

Form is responsible for keeping track of the menus on the form and passing on the relevant windows messages to the menu. Without this, popup events on the menu weren't firing. This PR copies in the relevant code from the older version of Winforms before MainMenu was removed.